requirements.txt
pypdf 
reportlab
vertexai
google-cloud-aiplatform
google-genai
openai
google-cloud-texttospeech
google-api-core
numpy
moviepy
playwright
imageio
imageio-ffmpeg